pyproject.toml
[project]
name = "defender-hunting-mcp"
version = "0.1.0"
description = "MCP Server for Microsoft Defender Advanced Hunting"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
